backend/requirements.txt
Django==3.0.8
psycopg2-binary==2.8.5
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Pillow==7.2.0
